0.3.0-beta.6
Browser version works again
