+++ Makefile
@@ -4,13 +4,15 @@ deps:
 	npm install
 
 test: deps
-	npm run test
-	
+	npm run test:node
+	npm run test:browser
+
 build: test
 	npm run build
+	cp dist/index.min.js examples/browser/lib/ipfs-browser-daemon.min.js
 
 clean:
 	rm -rf ipfs/
 	rm -rf node_modules/
 
-.PHONY: all deps test
+.PHONY: all test
